refactor(vehicle status): rename components
