+++ src/components/VehicleStatus/OpenVehicleStatus.vue
