+++ src/components/VehicleStatus/VehicleConnection.vue
