+++ src/components/VehicleStatus/VehicleCoordinate.vue
